# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Centros",
        "description": "Documentação da API de Centros, incluindo operações de CRUD e relacionamentos.",
        "version": "1.1.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API principal"
        }
    ],
    "paths": {
...
        "/horarios": {
            "get": {
                "tags": [
                    "Horários"
                ],
                "summary": "Listar todos os horários",
                "operationId": "8dcfbcbf2264d56095e7f5ea2dd36450",
                "responses": {
                    "200": {
                        "description": "Lista de horários",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Horario"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Horários"
                ],
                "summary": "Criar um novo horário",
                "operationId": "dcb0d05c43776ff8a494539fd4088b1c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HorarioInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Horário criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Horario"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/horarios/{id}": {
            "get": {
                "tags": [
                    "Horários"
                ],
                "summary": "Buscar horário por ID",
                "operationId": "4b59aaf0c35778d3f84c2003f8746c65",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Horário encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Horario"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Horário não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Horários"
                ],
                "summary": "Atualizar horário (não permite editar curso_id e centro_id)",
                "operationId": "5438beb885fe7049a2f6fbcbe9bbdcd1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HorarioUpdateInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Horário atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Horario"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Horário não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Horários"
                ],
                "summary": "Deletar horário",
                "operationId": "79039ed9290e3544523504ec4e85e7f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Horário deletado"
                    },
                    "404": {
                        "description": "Horário não encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Horario": {
                "required": [
                    "id",
                    "curso_id",
                    "centro_id",
                    "dia_semana",
                    "periodo",
                    "hora_inicio",
                    "hora_fim"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "curso_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "centro_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "dia_semana": {
                        "type": "string",
                        "example": "Segunda"
                    },
                    "periodo": {
                        "type": "string",
                        "example": "manhã"
                    },
                    "hora_inicio": {
                        "type": "string",
                        "example": "08:00"
                    },
                    "hora_fim": {
                        "type": "string",
                        "example": "10:00"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-31T08:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-31T08:00:00Z"
                    }
                },
                "type": "object",
                "example": {
                    "id": 1,
                    "curso_id": 2,
                    "centro_id": 3,
                    "dia_semana": "Segunda",
                    "periodo": "manhã",
                    "hora_inicio": "08:00",
                    "hora_fim": "10:00",
                    "created_at": "2025-07-31T08:00:00Z",
                    "updated_at": "2025-07-31T08:00:00Z"
                }
            },
            "HorarioInput": {
                "required": [
                    "curso_id",
                    "centro_id",
                    "dia_semana",
                    "periodo",
                    "hora_inicio",
                    "hora_fim"
                ],
                "properties": {
                    "curso_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "centro_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "dia_semana": {
                        "type": "string",
                        "example": "Segunda"
                    },
                    "periodo": {
                        "type": "string",
                        "example": "manhã"
                    },
                    "hora_inicio": {
                        "type": "string",
                        "example": "08:00"
                    },
                    "hora_fim": {
                        "type": "string",
                        "example": "10:00"
                    }
                },
                "type": "object",
                "example": {
                    "curso_id": 2,
                    "centro_id": 3,
                    "dia_semana": "Segunda",
                    "periodo": "manhã",
                    "hora_inicio": "08:00",
                    "hora_fim": "10:00"
                }
            },
            "HorarioUpdateInput": {
                "required": [
                    "dia_semana",
                    "periodo",
                    "hora_inicio",
                    "hora_fim"
                ],
                "properties": {
                    "dia_semana": {
                        "type": "string",
                        "example": "Segunda"
                    },
                    "periodo": {
                        "type": "string",
                        "example": "manhã"
                    },
                    "hora_inicio": {
                        "type": "string",
                        "example": "08:00"
                    },
                    "hora_fim": {
                        "type": "string",
                        "example": "10:00"
                    }
                },
                "type": "object",
                "example": {
                    "dia_semana": "Segunda",
                    "periodo": "manhã",
                    "hora_inicio": "08:00",
                    "hora_fim": "10:00"
                }
            },
...
        }
    },
...
}